refactor: move App to data/model package
