+++ app/src/main/java/dev/mslalith/focuslauncher/data/database/entities/AppRoom.kt
@@ -2,7 +2,7 @@ package dev.mslalith.focuslauncher.data.database.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import dev.mslalith.focuslauncher.data.App
+import dev.mslalith.focuslauncher.data.model.App
 import dev.mslalith.focuslauncher.utils.Constants
 
 @Entity(tableName = Constants.Database.APPS_TABLE_NAME)
