+++ app/src/main/java/dev/mslalith/focuslauncher/data/repository/AppDrawerRepo.kt
@@ -1,7 +1,7 @@
 package dev.mslalith.focuslauncher.data.repository
 
 import dev.mslalith.focuslauncher.data.database.dao.AppsDao
-import dev.mslalith.focuslauncher.data.App
+import dev.mslalith.focuslauncher.data.model.App
 import dev.mslalith.focuslauncher.data.database.entities.AppRoom
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
