+++ app/src/main/java/dev/mslalith/focuslauncher/data/repository/FavoritesRepo.kt
@@ -1,6 +1,6 @@
 package dev.mslalith.focuslauncher.data.repository
 
-import dev.mslalith.focuslauncher.data.App
+import dev.mslalith.focuslauncher.data.model.App
 import dev.mslalith.focuslauncher.data.database.dao.AppsDao
 import dev.mslalith.focuslauncher.data.database.dao.FavoriteAppsDao
 import dev.mslalith.focuslauncher.data.database.entities.FavoriteAppRoom
