+++ app/src/main/java/dev/mslalith/focuslauncher/extensions/UtilityExtensions.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Handler
 import android.os.Looper
-import dev.mslalith.focuslauncher.data.App
+import dev.mslalith.focuslauncher.data.model.App
 import dev.mslalith.focuslauncher.data.models.AppWithIcon
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
