+++ app/src/main/java/dev/mslalith/focuslauncher/ui/viewmodels/AppsViewModel.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.mslalith.focuslauncher.data.App
+import dev.mslalith.focuslauncher.data.model.App
 import dev.mslalith.focuslauncher.data.models.SelectedApp
 import dev.mslalith.focuslauncher.data.repository.AppDrawerRepo
 import dev.mslalith.focuslauncher.data.repository.FavoritesRepo
