+++ data/src/main/java/dev/mslalith/focuslauncher/data/model/App.kt
@@ -1,0 +1,7 @@
+package dev.mslalith.focuslauncher.data.model
+
+data class App(
+    val name: String,
+    val packageName: String,
+    val isSystem: Boolean,
+)
